third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEConvolveMatrixElement-dom-order-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'order' attribute of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS convolveMatrixElement.getAttribute('order') is "2 2"
PASS convolveMatrixElement.getAttribute('order') is "3 3"
PASS successfullyParsed is true

TEST COMPLETE
